# openapi.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "AdopteUnAnimal",
		"version": "1.0.0",
		"description": "Cette API REST développée avec [Express](https://expressjs.com/), permet de gérer des animaux à adopter. Vous pouvez manipuler des Pets, Users, SPA, Messages et Adoptions."
	},
	"servers": [
		{
			"url": "https://archioweb-adopteunanimal.onrender.com/"
		}
	],
...
	"paths": {
...
		"/pets/{id}/adoptions": {
			"get": {
				"tags": ["pets"],
				"summary": "Récupère les adoptions d'un animal par ID",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": { "type": "string" }
					}
				],
				"responses": {
					"200": {
						"description": "Liste des adoptions",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": { "$ref": "#/components/schemas/Adoption" }
								}
							}
						}
					}
				}
			}
		},
...
		"/users/{id}/adoptions": {
			"get": {
				"tags": ["users"],
				"summary": "Récupère les adoptions d'un utilisateur par ID",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": { "type": "string" }
					}
				],
				"responses": {
					"200": {
						"description": "Liste des adoptions",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": { "$ref": "#/components/schemas/Adoption" }
								}
							}
						}
					},
					"404": {
						"description": "Aucune adoption trouvée"
					}
				}
			}
		},
...
		"/adoptions": {
			"get": {
				"tags": ["adoptions"],
				"summary": "Liste toutes les adoptions",
				"responses": {
					"200": {
						"description": "Liste des adoptions",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Adoption"
									}
								}
							}
						}
					},
					"400": {
						"description": "Requête invalide"
					},
					"401": {
						"description": "Non autorisé"
					},
					"404": {
						"description": "Non trouvé"
					},
					"500": {
						"description": "Erreur interne du serveur"
					}
				}
			},
			"post": {
				"tags": ["adoptions"],
				"summary": "Ajoute une nouvelle adoption",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Adoption"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Adoption créée",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Adoption"
								}
							}
						}
					},
					"400": {
						"description": "Requête invalide"
					},
					"401": {
						"description": "Non autorisé"
					},
					"500": {
						"description": "Erreur interne du serveur"
					}
				}
			}
		},
		"/adoptions/{id}": {
			"get": {
				"tags": ["adoptions"],
				"summary": "Récupère une adoption par ID",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Détails de l'adoption",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Adoption"
								}
							}
						}
					},
					"400": {
						"description": "Requête invalide"
					},
					"401": {
						"description": "Non autorisé"
					},
					"404": {
						"description": "Non trouvé"
					},
					"500": {
						"description": "Erreur interne du serveur"
					}
				}
			},
			"delete": {
				"tags": ["adoptions"],
				"summary": "Supprime une adoption par ID",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Adoption supprimée"
					},
					"400": {
						"description": "Requête invalide"
					},
					"401": {
						"description": "Non autorisé"
					},
					"404": {
						"description": "Non trouvé"
					},
					"500": {
						"description": "Erreur interne du serveur"
					}
				}
			}
		},
		"/adoptions/{id}/messages": {
			"get": {
				"tags": ["adoptions"],
				"summary": "Liste tous les messages d'une adoption",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Liste des messages",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Message"
									}
								}
							}
						}
					},
					"400": {
						"description": "Requête invalide"
					},
					"401": {
						"description": "Non autorisé"
					},
					"404": {
						"description": "Non trouvé"
					},
					"500": {
						"description": "Erreur interne du serveur"
					}
				}
			},
			"post": {
				"tags": ["adoptions"],
				"summary": "Ajoute un nouveau message à une adoption",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Message"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Message créé",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Message"
								}
							}
						}
					},
					"400": {
						"description": "Requête invalide"
					},
					"401": {
						"description": "Non autorisé"
					},
					"500": {
						"description": "Erreur interne du serveur"
					}
				}
			}
		},
		"/adoptions/{id}/messages/{msg_id}": {
			"delete": {
				"tags": ["adoptions"],
				"summary": "Supprime un message par ID",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "msg_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Message supprimé"
					},
					"400": {
						"description": "Requête invalide"
					},
					"401": {
						"description": "Non autorisé"
					},
					"404": {
						"description": "Non trouvé"
					},
					"500": {
						"description": "Erreur interne du serveur"
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
			"Pet": {
				"type": "object",
				"properties": {
					"nom": { "type": "string" },
					"age": { "type": "number" },
					"description": { "type": "string" },
					"images": {
						"type": "array",
						"items": {
							"type": "object",
							"properties": {
								"data": { "type": "string", "format": "byte" },
								"imgType": { "type": "string" }
							}
						}
					},
					"tags": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Tag"
						}
					},
					"spa_id": {
						"$ref": "#/components/schemas/Spa"
					},
					"likes_count": {
						"type": "integer"
					},
					"dislikes_count": {
						"type": "integer"
					},
					"isAdopted": {
						"type": "boolean"
					}
				},
				"required": ["nom", "age", "description"]
			},
...
			"User": {
				"type": "object",
				"properties": {
					"_id": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"likes": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Pet"
						}
					},
					"dislikes": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Pet"
						}
					}
				},
				"required": ["firstName", "lastName", "email"]
			},
			"Tag": {
				"type": "object",
				"properties": {
					"_id": {
						"type": "string"
					},
					"nom": {
						"type": "string"
					}
				},
				"required": ["nom"]
			},
			"Spa": {
				"type": "object",
				"properties": {
					"_id": {
						"type": "string"
					},
					"nom": {
						"type": "string"
					},
					"adresse": {
						"type": "string"
					},
					"latitude": {
						"type": "string"
					},
					"longitude": {
						"type": "string"
					},
					"user_id": {
						"type": "string"
					}
				},
				"required": ["nom", "adresse", "latitude", "longitude"]
			},
			"Adoption": {
				"type": "object",
				"properties": {
					"_id": {
						"type": "string"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"user_id": {
						"$ref": "#/components/schemas/User"
					},
					"pet_id": {
						"$ref": "#/components/schemas/Pet"
					},
					"status": {
						"type": "string",
						"enum": ["pending", "accepted", "rejected"]
					},
					"messages": {
						"type": "array",
						"items": {
							"type": "object",
							"properties": {
								"_id": {
									"type": "string"
								},
								"content": {
									"type": "string"
								},
								"date": {
									"type": "string",
									"format": "date-time"
								},
								"user_id": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				}
			},
			"Message": {
				"type": "object",
				"properties": {
					"_id": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"user_id": {
						"type": "string"
					}
				}
			},
...
		}
	}
}